Very temp test.
All frameworks can work but:

- Everything is now `np` for ease of sanity.

Currently:

10s: PT
30s: TF
42s: Flax

```python
from transformers import FlaxRobertaModel, RobertaTokenizerFast,
RobertaForMaskedLM, pipeline
import numpy as np
import datetime
import tqdm
import random

def data(n):
    for _ in range(n):
        yield "JAX/Flax is amazing <mask>"

def flax(n):
    print("----")
    print("Flax")
    start = datetime.datetime.now()
    pipe = pipeline(
        model="distilbert-base-uncased-finetuned-sst-2-english",
        device=0,
        framework="flax",
        model_kwargs={"from_pt": True},
    )
    print("Loading flax", datetime.datetime.now() - start)

    for out in tqdm.tqdm(pipe(data(n), batch_size=512)):
        pass

def tf(n):
    print("----")
    print("TF")
    start = datetime.datetime.now()
    pipe =
pipeline(model="distilbert-base-uncased-finetuned-sst-2-english",
device=0, framework="tf")
    print("Loading TF", datetime.datetime.now() - start)

    for out in tqdm.tqdm(pipe(data(n), batch_size=512)):
        pass

def pt(n):
    print("----")
    print("PT")
    start = datetime.datetime.now()
    pipe =
pipeline(model="distilbert-base-uncased-finetuned-sst-2-english",
device=0, framework="pt")
    print("Loading PT", datetime.datetime.now() - start)

    for out in tqdm.tqdm(pipe(data(n), batch_size=512)):
        pass
        # print(out)

if __name__ == "__main__":
    n = 20000
    # pt(n)
    # flax(n)
    tf(n)
```
