+++ tests/test_pipelines_common.py
@@ -35,7 +35,7 @@ from transformers import (
 )
 from transformers.pipelines import get_task
 from transformers.pipelines.base import _pad
-from transformers.testing_utils import is_pipeline_test, nested_simplify, require_tf, require_torch
+from transformers.testing_utils import is_pipeline_test, nested_simplify, require_flax, require_tf, require_torch
 
 
 logger = logging.getLogger(__name__)
@@ -320,6 +320,25 @@ class CommonPipelineTest(unittest.TestCase):
 
         pipe = pipeline(model="hf-internal-testing/tiny-random-distilbert", framework="tf")
         out = pipe("This is a test")
+        self.assertEqual(nested_simplify(out), {"label": "LABEL_1", "score": 0.502})
+        pipe = pipeline(model="Narsil/tiny-distilbert-sequence-classification", framework="tf")
+        results = []
+        for out in pipe(data(10)):
+            self.assertEqual(nested_simplify(out), {"label": "LABEL_1", "score": 0.502})
+            results.append(out)
+        self.assertEqual(len(results), 10)
+
+    @require_flax
+    def test_iterator_data_flax(self):
+        def data(n: int):
+            for _ in range(n):
+                yield "This is a test"
+
+        pipe = pipeline(
+            model="Narsil/tiny-distilbert-sequence-classification", framework="flax", model_kwargs={"from_pt": True}
+        )
+        out = pipe("This is a test")
+        self.assertEqual(nested_simplify(out), {"label": "LABEL_1", "score": 0.502})
         results = []
         for out in pipe(data(10)):
             self.assertEqual(nested_simplify(out), {"label": "LABEL_0", "score": 0.504})
